AMAZING... modal working properly, edit/save functions working properly, just need to address user functionality so only the user can edit/delete their own articles
