+++ src/index.html
@@ -10,6 +10,7 @@
 </head>
 
 <body>
+  <div id="modalTarget"></div>
   <header><img src="../imgs/Icon.png" alt="plenty of witches logo" /></header>
   <article class="container">
     <article class="auth">
@@ -24,9 +25,9 @@
 
 
       <section class="articleContainer">
-        <div class="articleArea">
-          <section class="articleFormContainer"></section>
-          <section class="articleListContainer"></section>
+        <div class="articleArea"></div>
+        <section class="articleFormContainer"></section>
+        <section class="articleListContainer"></section>
         </div>
       </section>
