+++ src/scripts/articles/Article.js
@@ -7,7 +7,7 @@ export const ArticleHTMLConverter = (articleObject) => {
 		<section class="article">
 			<div class="article--timestamp">Date: ${new Date(articleObject.date).toLocaleDateString('en-US')}</div>
             <div class="article--title">News Title: ${articleObject.newsTitle}</div>
-            <div class="article--content">Synopsis: ${articleObject.newsSynopsis}</div>
+            <div class="article--content">Synopsis: ${articleObject.newsContent}</div>
             <div class="article--url">URL: ${articleObject.newsURL}</div>
             <button id="deleteArticle--${articleObject.id}" class="deleteArticleButton">Delete</button>
             <button id="editArticle--${articleObject.id}" class="editArticleButton">Edit</button>
